% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multintersect.R
\name{breakStrings}
\alias{breakStrings}
\title{breakStrings}
\usage{
breakStrings(x, minSizeForBreak = 20, lb = "\\n")
}
\arguments{
\item{x}{a character vector}

\item{minSizeForBreak}{the minimum number of characters to break on two lines (default 20)}

\item{lb}{the line break character (default line break)}
}
\value{
a character vector of length=length(x)
}
\description{
breaks a string of words (or vector thereof) into two lines
}
