% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multintersect.R
\name{dotplot.multintersect}
\alias{dotplot.multintersect}
\title{dotplot.multintersect}
\usage{
dotplot.multintersect(m, sizeRange = c(0, 20), colors = c(low = "blue",
  mid = "grey", high = "yellow"), forceDivergent = FALSE)
}
\arguments{
\item{m}{A list as produced by `multintersect`}

\item{sizeRange}{Size range of the dots}

\item{colors}{A named vector of 'low', 'mid', and 'high' colors}

\item{forceDivergent}{Logical; whether to force the use of a divergent color
palette; otherwise only low/high colors are used when there is no negative 
enrichment.}
}
\value{
A ggplot.
}
\description{
Plots the results of `multintersect` using a dot plot
}
