% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multintersect.R
\name{getEnrichment}
\alias{getEnrichment}
\title{getEnrichment}
\usage{
getEnrichment(set1, set2, universe)
}
\arguments{
\item{set1}{the set in which to test for enrichment}

\item{set2}{the set for which to test enrichment}

\item{universe}{either a character vector containing the universe/background (recommended), or an integer indicating the size of the universe}
}
\value{
A numeric value indicating the enrichment ratio (observed/expected)
}
\description{
Returns the enrichment ratio of set2 in set1
}
