% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multintersect.R
\name{multintersect}
\alias{multintersect}
\title{multintersect}
\usage{
multintersect(ll, ll2 = NULL, universe = NULL, addSetSize = TRUE,
  breakNames = NULL)
}
\arguments{
\item{ll}{A list of vectors to be compared}

\item{ll2}{An optional second list of vectors}

\item{universe}{An optional vector of the universe (all terms), used to calculate overlap probabilities and enrichments. If NULL, the union of all lists will be used. Only elements of `ll` and `ll2` that are in the `universe` will be considered.}

\item{addSetSize}{Logical; whether to add the set sizes to set names (default TRUE)}

\item{breakNames}{If not NULL (default), should be an number indicating the character length threshold above which set names should be split onto two lines.}
}
\value{
Returns a list
}
\description{
Performs pair-wise overlaps between the elements of one or two lists
}
