% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multintersect.R
\name{multintersect.test}
\alias{multintersect.test}
\title{multintersect.test}
\usage{
multintersect.test(ll, u, nperm = 2000, two.tailed = FALSE)
}
\arguments{
\item{ll}{A list of sets.}

\item{u}{The universe.}

\item{nperm}{The number of permutations to use.}

\item{two.tailed}{Whether the test should be two-tailed (defaults to testing
a greater alternative)}
}
\value{
A p-value, with precision roughly 1/nperm
}
\description{
Tests the significance of an overlap across multiple sets using 
permutations.
}
\examples{
# create random sets
ll <- lapply(1:3,FUN=function(x) sample(letters, 8))
multintersect.test(ll, u=letters)
}
