% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multintersect.R
\name{overlap.prob}
\alias{overlap.prob}
\title{overlap.prob}
\usage{
overlap.prob(set1, set2, universe, lower = F)
}
\arguments{
\item{set1}{a character vector}

\item{set2}{a character vector}

\item{universe}{either a character vector containing the universe/background (recommended), or an integer indicating the size of the universe}

\item{lower}{logical; whether the lower tail should be considered (default FALSE). By default the function tests for an enrichment, set to TRUE to test for a depletion.}
}
\value{
the probability of the overlap
}
\description{
Calculates the probability of the observed overlap between two sets on the basis of the hypergeometric distribution.
}
\examples{
overlap.prob( c("A","B","C","D"), c("B","D","G"), LETTERS )

}
