% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multintersect.R
\name{plot.multintersect}
\alias{plot.multintersect}
\title{plot.multintersect}
\usage{
\method{plot}{multintersect}(res, keyCol = "log2Enrichment",
  keyWrite = "overlap", margin = 100, title = "", cluster = TRUE)
}
\arguments{
\item{res}{A list as produced by `multintersect`}

\item{keyCol}{The values to use for the heatmap's colors. Possible values are: prob, enrichment, log2Enrichment, log10Prob, overlap, jaccard}

\item{keyWrite}{The values to be written in the cells. Possible values are: prob, enrichment, log2Enrichment, log10Prob, overlap, jaccard}

\item{margin}{Either a single number indicating the size of the left and bottom margins, or a list (l,r,b,t,pad) indicating the size of each margin.}

\item{title}{Plot title, default ''}

\item{cluster}{Logical; whether to cluster rows and columns (default TRUE)}
}
\value{
Plots a plotly heatmap
}
\description{
Plots the results of `multintersect` using an interactive heatmap
}
